fix(Jenkinsfile): Fix jenkins pipline for CI
It appears that the current stash and unstash has some issues with asynchronous behaviour.
For example an unstash occurs on a node in a working directory that is different than what is expected when running a test.
It also appears that some directories are not being cleaned.

The following commit makes a number of changes to fix that:

- makes jenkinsfile declarative as this is better supported
- run all tests on node before releasing to fix any shared workspace problems
- Cleanup function names and steps to make it more readable
- Add timeouts to overall process and timeout per node after it starts
- Handle errors if unstash fails only stop the node
- Allow a timeout/stop to exit the whole set of tests