dist/tools: Add better parser for finding env
This adds a env_parser python script that allows better control over how to show the env.
It allows a human readable version to be printed to the console and an xml to be generated.

This should allow metadata to be added to the ci test results.
The print_environment.sh is updated to use it but should be transparent the current ci system.