make: Add robot-plot target
This allows us to call the robot plot python script when the test has ROBOT_HAS_PLOT variable set.
If a plot is expected and not there it will error.
If a plot is not expected it will skip the call.

With this we can catch missing plots easier.
It also removes the plotting function from the robot-html target.